Include test files so test suite passes from PyPi
Package distributed from PyPi is missing the test data files so the test suite will always fail if the source is fetched from there. The test python files get pulled into the package because they have a .py extension, but the data files the test runs again are missed because they do not.

This commit grafts in test/ (pulls in everything) and then excludes pyc, pyo & pyd files per the recommended way of doing this in the python packaging manual.